texture manipulation

+++ source/private/shader.cpp
@@ -91,5 +91,5 @@ void Shader::setInt(const std::string& name, int value) const
 
 void Shader::setFloat(const std::string& name, float value) const
 {
-  glUniform1f(glGetUniformLocation(m_ID, name.c_str()), value);
+  glUniform1f(glGetUniformLocation(m_ID, name.c_str()), static_cast<float>(value));
 }
